fix some format error

+++ yunxiao/env_deploy_script/env_script/init_project_env.sh
@@ -2,7 +2,7 @@
 ###
 # @Author: robert zhang
 # @Date: 2019-09-02 12:23:30
- # @LastEditTime: 2020-08-29 20:28:22
+ # @LastEditTime: 2020-09-02 12:50:17
  # @LastEditors: robert zhang
 # @Description: 初始化执行脚本
 # @
@@ -11,10 +11,10 @@
 # $1:应用名称
 # $2:公共环境--环境类型
 
-source $HOME/env_script/commons
+source $HOME/env_script/functions
 
-APP_NAME=$1
-ENV_TYPE=$2
+export APP_NAME=$1
+export ENV_TYPE=$2
 #根据agent的日志配置项决定
 LOG_DIR=logs
 
@@ -22,7 +22,7 @@ LOG_DIR=logs
 clean_env
 
 # 创建log目录
-[ -d "$LOG_DIR" ] && rm -rf $LOG_DIR/*
+[ -d "$LOG_DIR" ] && rm -rf "${LOG_DIR:?}/*"
 
 mkdir $LOG_DIR
